server/requirements.txt
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-cors-headers==2.4.0
django-filter==2.1.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.1
idna==2.8
numpy==1.16.2
oauth2client==4.1.3
oauthlib==3.0.1
pandas==0.24.1
pdfkit==0.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyDrive==1.3.1
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
xlrd==1.2.0
